Refactor api packages
