+++ api/entrypoint.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 
 # Run APP
-uwsgi --ini api/uwsgi.ini --py-autoreload 1
+./river-api
