+++ api/errors/invalid_token.go
@@ -1,0 +1,10 @@
+package errors
+
+type InvalidTokenError struct {
+	Message    string
+	StatusCode int
+}
+
+func (err *InvalidTokenError) Error() string {
+	return err.Message
+}
