+++ api/mapping/delete.go
@@ -1,0 +1,11 @@
+package mapping
+
+import "fmt"
+
+func Delete(resourceID string, batchID string) error {
+	err := Rdb.Del(fmt.Sprintf("%s:%s", batchID, resourceID))
+	if err != nil {
+		return err.Err()
+	}
+	return nil
+}
