--- api/requirements-tests.txt
@@ -1,3 +1,0 @@
--r requirements.txt
-nose2==0.9.1
-nose2[coverage_plugin]==0.9.1
