--- api/requirements.txt
@@ -1,7 +1,0 @@
-confluent_kafka==1.3.0
-flask==1.0.2
-flask-cors==3.0.8
-fluent-logger==0.9.6
-requests==2.21.0
-uuid==1.30
-uWSGI==2.0.18
