+++ api/routes/preview/preview.go
@@ -1,10 +1,11 @@
-package api
+package preview
 
 import (
 	"bytes"
 	"encoding/json"
-	"errors"
 	"fmt"
+	"github.com/arkhn/fhir-river/api/errors"
+	"github.com/arkhn/fhir-river/api/mapping"
 	"io/ioutil"
 	"net/http"
 
@@ -13,7 +14,7 @@ import (
 )
 
 // PreviewRequest is the body of the POST /preview request.
-type PreviewRequest struct {
+type Request struct {
 	// PrimaryKeyValues can be a list of strings (eg ["E65"]) or a list of integers ([59])
 	PrimaryKeyValues []interface{} `json:"primary_key_values"`
 	ResourceID       string        `json:"resource_id"`
@@ -50,7 +51,7 @@ func transform(resourceID, previewID string, rows []interface{}) (res []byte, er
 		// If everything went well, we go on
 	default:
 		// Return other errors
-		return nil, errors.New(string(body))
+		return nil, fmt.Errorf(string(body))
 	}
 
 	return body, nil
@@ -58,7 +59,7 @@ func transform(resourceID, previewID string, rows []interface{}) (res []byte, er
 
 // transform sends an HTTP request to the transformer service
 // using the PreviewRequest as JSON body. It returns the extrcted rows.
-func extract(preview *PreviewRequest) (rows []interface{}, err error) {
+func extract(preview *Request) (rows []interface{}, err error) {
 	jBody, _ := json.Marshal(preview)
 	url := fmt.Sprintf("%s/extract", extractorURL)
 
@@ -82,7 +83,7 @@ func extract(preview *PreviewRequest) (rows []interface{}, err error) {
 		if err != nil {
 			return nil, err
 		}
-		return nil, errors.New(string(body))
+		return nil, fmt.Errorf(string(body))
 	}
 
 	body := struct {
@@ -101,7 +102,7 @@ func extract(preview *PreviewRequest) (rows []interface{}, err error) {
 // the FHIR transformation of a set of rows for a given resource.
 func Preview(w http.ResponseWriter, r *http.Request) {
 	// decode the request body
-	body := PreviewRequest{}
+	body := Request{}
 	err := json.NewDecoder(r.Body).Decode(&body)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusBadRequest)
@@ -121,11 +122,11 @@ func Preview(w http.ResponseWriter, r *http.Request) {
 	authorizationHeader := r.Header.Get("Authorization")
 
 	// Fetch and store the mappings to use for the preview
-	resourceMapping, err := fetchMapping(body.ResourceID, authorizationHeader)
+	resourceMapping, err := mapping.Fetch(body.ResourceID, authorizationHeader)
 	if err != nil {
 		switch e := err.(type) {
-		case *invalidTokenError:
-			http.Error(w, err.Error(), e.statusCode)
+		case *errors.InvalidTokenError:
+			http.Error(w, err.Error(), e.StatusCode)
 		default:
 			http.Error(w, err.Error(), http.StatusBadRequest)
 		}
@@ -139,7 +140,7 @@ func Preview(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// cache the mapping in redis
-	err = storeMapping(serializedMapping, body.ResourceID, body.PreviewID)
+	err = mapping.Store(serializedMapping, body.ResourceID, body.PreviewID)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
@@ -160,12 +161,12 @@ func Preview(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// delete the mapping from redis
-	err = deleteMapping(body.ResourceID, body.PreviewID)
+	err = mapping.Delete(body.ResourceID, body.PreviewID)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
 
 	// return the response of the transformer service
-	fmt.Fprint(w, string(res))
+	_, _ = fmt.Fprint(w, string(res))
 }
