--- api/uwsgi.ini
@@ -1,13 +1,0 @@
-[uwsgi]
-
-module = api.src.app:app
-master = true
-
-processes = 1
-threads = 4
-
-http = 0.0.0.0:3000
-
-vacuum = true
-
-die-on-term = true
