--- apigo/api/errors.go
@@ -1,10 +1,0 @@
-package api
-
-type invalidTokenError struct {
-	message    string
-	statusCode int
-}
-
-func (err *invalidTokenError) Error() string {
-	return err.message
-}
