--- apigo/entrypoint.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-
-# Run APP
-./river-api
